upgrade sources; add 'keep answers' feature; small tuning

+++ run.sh
@@ -19,8 +19,11 @@ main() {
 
     echo "Clone repo..."
     git clone $REPO $DIR/interview-tasks
-    echo "Cleanup answers..."
-    rm -rf $DIR/interview-tasks/**/answers
+
+    if [[ $KEEP_ANSWERS != yes ]]; then
+        echo "Cleanup answers..."
+        rm -rf $DIR/interview-tasks/**/answers
+    fi
 
     if [[ ! -z $TEST ]]; then
         IFS=','
@@ -58,6 +61,9 @@ main() {
             cp -r $DIR/interview-tasks/$val $DIR/project/
         done
     fi
+
+    rm $DIR/project/Dockerfile
+    rm $DIR/project/run.sh
 
     echo "Generate password..."
     PASS=`pwgen 16 1`
@@ -108,7 +114,7 @@ cert: false" > $DIR/.config/code-server/config.yaml
 
     i=0
     while [[ ${STATUS_RECEIVED} != 200 ]]; do
-        if [[ $i -ge 180 ]]; then
+        if [[ $i -ge 600 ]]; then
             echo "Failed to get code-server ready for $i seconds"
             exit -1
         fi
@@ -136,12 +142,13 @@ cert: false" > $DIR/.config/code-server/config.yaml
         PUBLIC_URL="$PUBLIC_URL:$CODE_SERVER_PORT"
     fi
 
-    SQL_FILES=`find $DIR -wholename "$DIR/project/**/init/*.sql" | sed -e "s|$DIR/||g"`
+    SQL_FILES=`docker exec $CONTAINER find . -type f -name *.sql | sed -e 's|^\./||g'`
 
     if ! [[ -z $SQL_FILES ]]; then
         docker exec $CONTAINER sudo service postgresql start
         for file in $SQL_FILES; do
-            docker exec $CONTAINER psql -U postgres -f "/home/coder/$file"
+            echo $file
+            docker exec $CONTAINER psql -U postgres -f $file
         done
     fi
